feat(frontend): persistent AppShell so editor↔library keeps the chrome
The editor and library are already one SPA, but App.vue was a bare
<router-view> mapping /editor and /library to two independent top-level
components. Each tab switch unmounted one whole tree and mounted the other,
rebuilding the shared chrome (toolbars, tab-bar, settings + traject menus) —
which reads as a full refresh even though the document never reloaded.

Introduce AppShell.vue as a persistent parent route holding the shared chrome
and a nested <router-view>; EditorView/LibraryView (renamed from *App.vue) are
now its children, so a switch swaps only the nested view and the shell instance
is reused — no chrome rebuild.

Per-view toolbar extras are published through a small module-level chrome store
(useAppChrome.js) rather than <Teleport>: nldd-toolbar only projects direct-
child slotted items, so teleporting would reorder the bar. The shell renders
every item in the original order; the active view registers its SearchPopover
and (editor only) its PR indicator + document tabs, cleared on unmount.

Route names, paths and redirects are unchanged, so deep links and the
traject-aware sectionTarget logic keep working. The shared settings menu now
lists the full 7-flag editor set (the library previously showed only 4).

Verified: 268 unit tests pass, production build splits EditorView lazily, and a
Playwright round-trip confirms the same shell DOM node survives both switch
directions while the document-tabs pane appears/clears with the editor.
